refactor(dashboard): split UsageDashboardView into shell + page sections
`UsageDashboardView.swift` carried the window controller, four navigation
types, the split-view shell, and all seven content pages in one 1097-line
file. Split it along the seams that were already there:

- `DashboardNavigation.swift` — window controller, `DashboardSection`,
  `SettingsSection`, `EnabledQuotaSourceCounter`, `DashboardNavigationStore`
- `DashboardOverviewSection` / `DashboardCostsSection` /
  `DashboardStatusSection` / `DashboardDiagnosticsSection` /
  `DashboardShareSection` — one page each, taking explicit inputs

Child views rather than cross-file extensions: `private` is file-scoped, so
extension-splitting would have forced all 15 `@StateObject` stores to
internal. Taking inputs explicitly matches how `ProviderStatusCard` and
`ProviderStatusTable` are already shared with the popover.

Three parallel section->refresh switches collapse into a single
`DashboardSection.refreshTarget` (+ `refreshesApiUsage`) mapping, and the
Automation feature gate moves into `SettingsSection.available(_:)`. Refresh
semantics are unchanged: `.costs` still refreshes API usage, `.share` and
`.optimize` still only rescan costs.

One behavior change, documented in the file: the shell used to kick
`runDiagnostics()` from `onChange(of: selectedSection)` while the page also
carried its own `.task(id:)`, so opening Diagnostics ran two concurrent
readiness sweeps over the keychain, filesystem, and SQLite — the `.task`
path bypassed the in-flight guard. The page now owns the single entry point.
No guard is needed there: `.task(id:)` starts the new run before the
cancelled one's `defer` clears the flag, so a guard would make the new run
bail; overlap is prevented by the Re-run button's `.disabled`.

Adds 18 tests covering the moved navigation API, each section's constructor,
and the lifted string helpers.

Shell: 1097 -> 369 lines.
